chore(sonar): resolve SonarCloud findings on PR #890
Fix the one gate-blocking reliability finding and the non-blocking code
smells the branch introduced.

- S6466 (bug): the realized-health warning sliced reasons[:5]; a slice
  cannot IndexError, but Sonar's dataflow flags it. Use itertools.islice,
  which reads as "first five" and clears new_reliability_rating.
- S104: split seven oversized modules along cohesive, cycle-free seams
  into siblings (snapshot_models, backend_host_inventory,
  _raes_apply_helpers, _compose_profile_index, _raes_conformance_probe,
  _raes_image_policy, _live_gate_variation); all now under 500 lines.
- S1142: reduce every >3-return function to <=3 via guard merges/helpers.
- FunctionComplexity: extract helpers in the artifact-availability and
  compose-profile-index loaders.
- S1172: mark interface-mandated unused params with del (repo convention).
- S1128: unquote forward-ref annotations / move a type-only import under
  TYPE_CHECKING so the imports read as used.
- S1309: drop inert `# noqa: BLE001` (ruff enforces only C901); keep the
  rationale as a leading comment.
- S139: exclude TYPE_CHECKING blocks from coverage globally instead of
  inline pragmas; move the remaining trailing comments; drop a dead branch.
- S2325: make the offline-gate stub's stateless methods static.
- S5778: hoist non-asserting setup out of pytest.raises blocks.

Re-exports preserve every external import and test patch target, and the
DeploymentBackend protocol keeps all methods via inheritance.
