+++ src/aptl/backends/raes.py
@@ -11,6 +11,10 @@ from raes_runtime.manager import RuntimeManager
 from raes_runtime.registry import RuntimeTarget
 from raes import SDLError, SDLInstantiationError, parse_sdl_file
 
+from aptl.backends._raes_apply_helpers import (
+    _drive_orchestrator_workflows,
+    _with_backend_failure_diagnostics,
+)
 from aptl.backends.raes_diagnostics import (
     render_raes_diagnostics,
 )
@@ -45,7 +49,6 @@ from aptl.utils.logging import get_logger
 from aptl.utils.redaction import redact
 
 if TYPE_CHECKING:
-    from raes_contracts.diagnostics import Diagnostic
     from raes_processor.models import ExecutionPlan
 
     from aptl.core.deployment.backend import DeploymentBackend
@@ -433,35 +436,6 @@ def _run_execution_plan(
     )
 
 
-def _with_backend_failure_diagnostics(
-    target: RuntimeTarget,
-    diagnostics: list["Diagnostic"],
-) -> list["Diagnostic"]:
-    """Re-attach the provisioner's own failure report to a failed apply.
-
-    ``raes_runtime``'s backend-call boundary replaces a failed backend apply's
-    diagnostics with its snapshot-contract / SEM-218 gate output — the gate
-    evaluates the never-realized snapshot, so every exact declaration reads as
-    unrealized. That hides the backend's actionable failure (for example a
-    Docker subnet conflict with its remediation steps) and severs the
-    retryable-code signal the SOC retry keys on. The backend's own report is
-    authoritative for what failed, so it renders first.
-    """
-    captured = tuple(getattr(target.provisioner, "last_failure_diagnostics", ()))
-    if not captured:
-        return diagnostics
-    seen = {
-        (diagnostic.code, diagnostic.address, diagnostic.message)
-        for diagnostic in diagnostics
-    }
-    fresh = [
-        diagnostic
-        for diagnostic in captured
-        if (diagnostic.code, diagnostic.address, diagnostic.message) not in seen
-    ]
-    return [*fresh, *diagnostics]
-
-
 def _apply_execution_plan(
     target: RuntimeTarget,
     execution_plan: "ExecutionPlan",
@@ -512,28 +486,3 @@ def _apply_execution_plan(
         run_id=run_id,
     )
     return failure, snapshot, False
-
-
-def _drive_orchestrator_workflows(
-    orchestrator: object,
-    evaluation_results: dict[str, dict[str, object]],
-    *,
-    run_store: RunStorageBackend | None = None,
-    run_id: str | None = None,
-) -> LabResult | None:
-    """Drive registered workflows and convert diagnostics to a lab failure."""
-
-    drive_diagnostics = []
-    if isinstance(orchestrator, AptlOrchestrator) and orchestrator.results():
-        drive_diagnostics = orchestrator.drive_workflows(
-            evaluation_results=evaluation_results,
-            run_store=run_store,
-            run_id=run_id,
-        )
-    failure = None
-    if drive_diagnostics:
-        failure = LabResult(
-            success=False,
-            error=render_raes_diagnostics(drive_diagnostics),
-        )
-    return failure
