+++ src/aptl/backends/raes_realization.py
@@ -44,6 +44,10 @@ from aptl.backends.raes_realization_model import (
     NodeRealization,
     _single_or_none,
 )
+from aptl.backends._raes_conformance_probe import (
+    _conformance_probe_services,
+    _is_raes_conformance_probe_node,
+)
 from aptl.backends.raes_realization_values import (
     mapping as _mapping,
     network_names as _network_names,
@@ -51,7 +55,6 @@ from aptl.backends.raes_realization_values import (
     optional_string as _optional_string,
     published_ports as _published_ports,
     resource_name as _resource_name,
-    service_names as _service_names,
     service_ports as _service_ports,
     static_address_assignments as _static_address_assignments,
     static_addresses as _static_addresses,
@@ -411,70 +414,6 @@ def _node_runtime(node_spec: Mapping[str, Any] | None) -> RuntimeConfiguration |
         return RuntimeConfiguration.model_validate(dict(raw))
     except (ValueError, TypeError):
         return None
-
-
-def _is_raes_conformance_probe_node(
-    resource: PlannedResource,
-    payload: Mapping[str, Any],
-) -> bool:
-    """Return whether a node is RAES' backend-neutral live probe."""
-
-    spec = _mapping(payload.get("spec"))
-    node_spec = _mapping(spec.get("node")) if spec else None
-    infra_spec = _mapping(spec.get("infrastructure")) if spec else None
-    return (
-        _has_raes_conformance_probe_identity(resource, payload)
-        and _has_empty_raes_probe_node_spec(node_spec)
-        and _has_empty_raes_probe_infra_spec(infra_spec)
-    )
-
-
-def _has_raes_conformance_probe_identity(
-    resource: PlannedResource,
-    payload: Mapping[str, Any],
-) -> bool:
-    """Return whether resource identity matches RAES' generic VM probe."""
-
-    return (
-        resource.address,
-        str(payload.get("name", "")),
-        str(payload.get("node_name", "")),
-        str(payload.get("node_type", "")),
-        str(payload.get("os_family", "")),
-    ) == ("provision.node.vm", "vm", "vm", "vm", "linux")
-
-
-def _has_empty_raes_probe_node_spec(
-    node_spec: Mapping[str, Any] | None,
-) -> bool:
-    """Return whether the generic probe has no concrete service source."""
-
-    return (
-        bool(node_spec)
-        and node_spec.get("source") is None
-        and not _service_names(node_spec)
-    )
-
-
-def _has_empty_raes_probe_infra_spec(
-    infra_spec: Mapping[str, Any] | None,
-) -> bool:
-    """Return whether the generic probe has no scenario network intent."""
-
-    return not _network_names(infra_spec) and not _static_addresses(infra_spec)
-
-
-def _conformance_probe_services(
-    profile_index: ComposeProfileIndex,
-    config: AptlConfig,
-) -> frozenset[str]:
-    """Bind RAES' generic probe to one enabled APTL service, if available."""
-
-    for profile in public_start_profiles(config):
-        for service_name, service in sorted(profile_index.services.items()):
-            if profile in service.profiles:
-                return frozenset({service_name})
-    return frozenset()
 
 
 def _container_name(
